requirements.txt
cloudscraper
requests
feedparser
pyTelegramBotAPI
deep-translator
google-generativeai